fsck.fat: Adhere to the fsck exit codes
fsck.fat is used as a filesystem-specific checker for the `fsck`. This
also causes `fsck` to return the same exit-codes given by `fsck.fat`.

In most cases this is already the case. One exception to that comes when
checking a read-only filesystem. In that case `fsck.fat` will return 6,
which for `fsck` means "Fiesystem errors left uncorrected" and "System
should reboot". When a more proper response would be to return 4,
"Filesystem errors uncorrected".

This commit solves that problem by introducing a new header file which
standardizes the exit codes used by `fsck.fat`.

Signed-off-by: Ricardo Ungerer <ungerer.ricardo@gmail.com>